+++ docs/source/News.rst
@@ -11,6 +11,8 @@ This document describes the releases of :mod:`repoze.what.plugins.sql`.
 =================================================
 
 * Synchronized to :mod:`repoze.what` v1.0.1 and its ``credentials`` dict.
+* Loaded the user object into the ``identity`` thanks to
+  :class:`repoze.who.plugins.sa.SQLAlchemyUserMDPlugin`.
 
 
 .. _repoze.what-sql-1.0a5: